requirements.txt
Django
Pillow
python-decouple
drf-yasg
djangorestframework
django-filter
psycopg2-binary
gunicorn
whitenoise
djangorestframework-simplejwt
social-auth-app-django